Change Budget type from int to decimal

+++ src/Won.Shared/Dtos/TripDto.cs
@@ -8,7 +8,7 @@
     public DateTime EndDate { get; set; }
     public string Location { get; set; } = string.Empty;
     public string Details { get; set; } = string.Empty;
-    public int Budget { get; set; }
+    public decimal Budget { get; set; }
     public int GroupSize { get; set; }
     }
 }
